Move particles to new package
