+++ src/main/java/com/flansmod/client/particle/EntityFMTracer.java
@@ -1,4 +1,4 @@
-package com.flansmod.client;
+package com.flansmod.client.particle;
 
 import org.lwjgl.opengl.GL11;
 
@@ -10,10 +10,10 @@ import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.World;
 
-public class EntityFMTracerRed extends EntityFX
+public class EntityFMTracer extends EntityFX
 {
-	public static ResourceLocation icon = new ResourceLocation("flansmod", "particle/FMTracerRed.png");
-	public EntityFMTracerRed(World w, double px, double py, double pz, double mx, double my, double mz)
+	public static ResourceLocation icon = new ResourceLocation("flansmod", "particle/FMTracer.png");
+	public EntityFMTracer(World w, double px, double py, double pz, double mx, double my, double mz)
 	{
 		super(w, px, py, pz, mx, my, mz);
 		this.particleMaxAge = 6;
@@ -22,13 +22,13 @@ public class EntityFMTracerRed extends EntityFX
 		this.motionX = mx;
 		this.motionY = my;
 		this.motionZ = mz;
-		FlansMod.proxy.spawnParticle("flansmod.FMTracerRed",
+		FlansMod.proxy.spawnParticle("flansmod.fmsmoke",
 				this.posX,
 				this.posY,
 				this.posZ,
 				0,0,0);
 
-		icon = new ResourceLocation("flansmod", "particle/FMTracerRed.png");
+		icon = new ResourceLocation("flansmod", "particle/FMTracer.png");
 	}
 	
 	public int getFXLayer()
@@ -57,14 +57,14 @@ public class EntityFMTracerRed extends EntityFX
 		int dstBlend = GL11.glGetInteger(GL11.GL_BLEND_DST);
 		//GL11.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
 		GL11.glDepthMask(false); 
-    	FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "particle/FMTracerRed.png"));
+    	FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "particle/FMTracer.png"));
 
         float scale = 0.6F - ((this.particleAge)*0.1F);
         float xPos = (float) (this.prevPosX + (this.posX - this.prevPosX) * (double) par2 - interpPosX);
         float yPos = (float) (this.prevPosY + (this.posY - this.prevPosY) * (double) par2 - interpPosY);
         float zPos = (float) (this.prevPosZ + (this.posZ - this.prevPosZ) * (double) par2 - interpPosZ);
         float colorIntensity = 1F;
-        //par1Tessellator.setColorOpaque_F(this.particleRed * colorIntensity, this.particleRed * colorIntensity, this.particleBlue * colorIntensity);//, 1.0F);
+        //par1Tessellator.setColorOpaque_F(this.particleRed * colorIntensity, this.particleGreen * colorIntensity, this.particleBlue * colorIntensity);//, 1.0F);
         par1Tessellator.setColorRGBA_F(this.particleRed * colorIntensity, this.particleGreen * (colorIntensity - this.particleAge*0.2F), this.particleBlue * (colorIntensity - this.particleAge*0.2F), (1F - this.particleAge*0.1F));
         par1Tessellator.addVertexWithUV((double) (xPos - par3 * scale - par6 * scale), (double) (yPos - par4 * scale), (double) (zPos - par5 * scale - par7 * scale), 0D, 1D);
         par1Tessellator.addVertexWithUV((double) (xPos - par3 * scale + par6 * scale), (double) (yPos + par4 * scale), (double) (zPos - par5 * scale + par7 * scale), 1D, 1D);
