+++ src/main/java/com/flansmod/client/particle/EntityFlash.java
@@ -1,39 +1,27 @@
-package com.flansmod.client;
-
-import java.util.List;
+package com.flansmod.client.particle;
 
 import org.lwjgl.opengl.GL11;
 
 import com.flansmod.common.FlansMod;
-import com.flansmod.common.teams.TeamsManager;
 
 import cpw.mods.fml.client.FMLClientHandler;
-import net.minecraft.block.Block;
 import net.minecraft.client.particle.EntityFX;
 import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityLiving;
-import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Blocks;
-import net.minecraft.util.AxisAlignedBB;
-import net.minecraft.util.DamageSource;
-import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.World;
 
-public class EntitySmokeGrenade extends EntityFX
+public class EntityFlash extends EntityFX
 {
-	public int dischargeTime;
-	public EntitySmokeGrenade(World w, double px, double py, double pz, double mx, double my, double mz)
+	public static ResourceLocation icon = new ResourceLocation("flansmod", "particle/Boom1.png");
+	public EntityFlash(World w, double px, double py, double pz, double mx, double my, double mz)
 	{
 		super(w, px, py, pz, mx, my, mz);
-		this.particleMaxAge *= 20;
+		this.particleMaxAge = 6;
 		this.particleGravity = 1;
 		this.motionX = mx;
 		this.motionY = my;
 		this.motionZ = mz;
-		this.dischargeTime = 20;
+		icon = new ResourceLocation("flansmod", "particle/Boom1.png");
 	}
 	
 	public int getFXLayer()
@@ -46,10 +34,15 @@ public class EntitySmokeGrenade extends EntityFX
 			return 1.0F;
 	}
 	
+	public int getBrightnessForRender(float par1)
+	{
+		return 15728880;
+	}
+	
     public void renderParticle(Tessellator par1Tessellator, float par2, float par3, float par4, float par5, float par6, float par7)
     {
-        //func_98187_b() = bindTexture();
     	GL11.glPushMatrix();
+        //func_98187_b() = bindTexture();
     	par1Tessellator.startDrawingQuads();
 		GL11.glAlphaFunc(GL11.GL_GREATER, 0.001F);
 		GL11.glEnable(GL11.GL_BLEND);
@@ -57,9 +50,9 @@ public class EntitySmokeGrenade extends EntityFX
 		int dstBlend = GL11.glGetInteger(GL11.GL_BLEND_DST);
 		GL11.glBlendFunc(1, GL11.GL_ONE_MINUS_SRC_ALPHA);
 		GL11.glDepthMask(false); 
-    	FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "particle/case.png"));
+    	FMLClientHandler.instance().getClient().renderEngine.bindTexture(icon);
 
-        float scale = 0.1F * this.particleScale;
+        float scale = 1F;
         float xPos = (float) (this.prevPosX + (this.posX - this.prevPosX) * (double) par2 - interpPosX);
         float yPos = (float) (this.prevPosY + (this.posY - this.prevPosY) * (double) par2 - interpPosY);
         float zPos = (float) (this.prevPosZ + (this.posZ - this.prevPosZ) * (double) par2 - interpPosZ);
@@ -76,13 +69,6 @@ public class EntitySmokeGrenade extends EntityFX
 		GL11.glDepthMask(true); 
 		GL11.glPopMatrix();
     }
-	
-	@Override
-    public AxisAlignedBB getBoundingBox()
-    {
-		boundingBox.expand(1, 1, 1);
-        return boundingBox;
-    }
 
 	public void onUpdate()
 	{
@@ -94,58 +80,19 @@ public class EntitySmokeGrenade extends EntityFX
 		{
 			this.setDead();
 		}
-		
-
-
-		this.motionY -= 0.04D * (double) this.particleGravity;
-		this.moveEntity(this.motionX, this.motionY, this.motionZ);
-		this.motionX *= 0.99;
-		this.motionY *= 0.99;
-		this.motionZ *= 0.99;
-		
-		if(worldObj.isAnyLiquid(this.boundingBox))
-		{
-			this.motionY = 1;
-			
-		}
-		
-		dischargeTime --;
-		
-		if(dischargeTime < 0)
-		{
-			double dx = (this.posX-this.prevPosX);
-			double dy = (this.posY-this.prevPosY);
-			double dz = (this.posZ-this.prevPosZ);
-			FlansMod.proxy.spawnParticle("flansmod.smokeburst",
-					this.posX,
-					this.posY,
-					this.posZ,
-					0,0,0);
-			
-			FlansMod.proxy.spawnParticle("flansmod.bigsmoke",
-					this.posX,
-					this.posY,
-					this.posZ,
-					0,0,0);
-			setDead();	
-		}
-		
-		int NUM = 5;
-		for(int i=0; i<NUM; i++)
-		{
-			double dx = (this.posX-this.prevPosX) / NUM;
-			double dy = (this.posY-this.prevPosY) / NUM;
-			double dz = (this.posZ-this.prevPosZ) / NUM;
-			FlansMod.proxy.spawnParticle("explode",
-				this.prevPosX + dx*i,
-				this.prevPosY + dy*i,
-				this.prevPosZ + dz*i,
-				0,0,0);
-		}
 
 		if(this.onGround)
 		{
 			setDead();
 		}
+		
+		if(this.particleAge == 1)icon = new ResourceLocation("flansmod", "particle/Boom2.png");
+		if(this.particleAge == 2)icon = new ResourceLocation("flansmod", "particle/Boom3.png");
+		if(this.particleAge == 3)icon = new ResourceLocation("flansmod", "particle/Boom4.png");
+		if(this.particleAge == 4)icon = new ResourceLocation("flansmod", "particle/Boom5.png");
+		if(this.particleAge == 5)icon = new ResourceLocation("flansmod", "particle/Boom6.png");
+		
+    	FMLClientHandler.instance().getClient().renderEngine.bindTexture(icon);
+
 	}
 }
