+++ src/main/java/com/flansmod/client/particle/EntityRocketexhaust.java
@@ -1,38 +1,27 @@
-package com.flansmod.client;
-
-import java.util.List;
+package com.flansmod.client.particle;
 
 import org.lwjgl.opengl.GL11;
 
 import com.flansmod.common.FlansMod;
-import com.flansmod.common.teams.TeamsManager;
 
 import cpw.mods.fml.client.FMLClientHandler;
-import net.minecraft.block.Block;
 import net.minecraft.client.particle.EntityFX;
 import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityLiving;
-import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Blocks;
-import net.minecraft.util.AxisAlignedBB;
-import net.minecraft.util.DamageSource;
-import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.World;
 
-public class EntityFlare extends EntityFX
+public class EntityRocketexhaust extends EntityFX
 {
-	public float damage = 5;
-	public EntityFlare(World w, double px, double py, double pz, double mx, double my, double mz)
+	public static ResourceLocation icon = new ResourceLocation("flansmod", "particle/FMSmoke.png");
+	public EntityRocketexhaust(World w, double px, double py, double pz, double mx, double my, double mz)
 	{
 		super(w, px, py, pz, mx, my, mz);
-		this.particleMaxAge *= 100;
+		this.particleMaxAge = 16;
 		this.particleGravity = 1;
 		this.motionX = mx;
 		this.motionY = my;
 		this.motionZ = mz;
+		icon = new ResourceLocation("flansmod", "particle/FMSmoke.png");
 	}
 	
 	public int getFXLayer()
@@ -45,31 +34,27 @@ public class EntityFlare extends EntityFX
 			return 1.0F;
 	}
 	
-	public int getBrightnessForRender(float par1)
-	{
-		return 15728880;
-	}
 	
     public void renderParticle(Tessellator par1Tessellator, float par2, float par3, float par4, float par5, float par6, float par7)
     {
         //func_98187_b() = bindTexture();
     	GL11.glPushMatrix();
-   	 	par1Tessellator.startDrawingQuads();
+    	par1Tessellator.startDrawingQuads();
 		GL11.glAlphaFunc(GL11.GL_GREATER, 0.001F);
 		GL11.glEnable(GL11.GL_BLEND);
 		int srcBlend = GL11.glGetInteger(GL11.GL_BLEND_SRC);
 		int dstBlend = GL11.glGetInteger(GL11.GL_BLEND_DST);
-		GL11.glBlendFunc(1, GL11.GL_ONE_MINUS_SRC_ALPHA);
+		GL11.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
 		GL11.glDepthMask(false); 
-    	FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "gui/flare.png"));
+    	FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "particle/FMSmoke.png"));
 
-        float scale = 0.3F * this.particleScale;
+        float scale = 10F + (this.particleAge * 0.05F);
         float xPos = (float) (this.prevPosX + (this.posX - this.prevPosX) * (double) par2 - interpPosX);
         float yPos = (float) (this.prevPosY + (this.posY - this.prevPosY) * (double) par2 - interpPosY);
         float zPos = (float) (this.prevPosZ + (this.posZ - this.prevPosZ) * (double) par2 - interpPosZ);
-        float colorIntensity = 1.0F;
-        par1Tessellator.setColorOpaque_F(this.particleRed * colorIntensity, this.particleGreen * colorIntensity, this.particleBlue * colorIntensity);//, 1.0F);
-
+        float colorIntensity = 1F;
+        //par1Tessellator.setColorOpaque_F(this.particleRed * colorIntensity, this.particleGreen * colorIntensity, this.particleBlue * colorIntensity);//, 1.0F);
+        par1Tessellator.setColorRGBA_F(this.particleRed * colorIntensity, this.particleGreen * colorIntensity, this.particleBlue * colorIntensity, (1F - this.particleAge*0.1F));
         par1Tessellator.addVertexWithUV((double) (xPos - par3 * scale - par6 * scale), (double) (yPos - par4 * scale), (double) (zPos - par5 * scale - par7 * scale), 0D, 1D);
         par1Tessellator.addVertexWithUV((double) (xPos - par3 * scale + par6 * scale), (double) (yPos + par4 * scale), (double) (zPos - par5 * scale + par7 * scale), 1D, 1D);
         par1Tessellator.addVertexWithUV((double) (xPos + par3 * scale + par6 * scale), (double) (yPos + par4 * scale), (double) (zPos + par5 * scale + par7 * scale), 1D, 0D);
@@ -79,13 +64,6 @@ public class EntityFlare extends EntityFX
 		GL11.glDisable(GL11.GL_BLEND);
 		GL11.glDepthMask(true); 
 		GL11.glPopMatrix();
-    }
-	
-	@Override
-    public AxisAlignedBB getBoundingBox()
-    {
-		boundingBox.expand(1, 1, 1);
-        return boundingBox;
     }
 
 	public void onUpdate()
@@ -98,31 +76,8 @@ public class EntityFlare extends EntityFX
 		{
 			this.setDead();
 		}
-
-		this.motionY -= 0.04D * (double) this.particleGravity;
+		this.motionY += 0.01;
 		this.moveEntity(this.motionX, this.motionY, this.motionZ);
-		this.motionX *= 0.99;
-		this.motionY *= 0.99;
-		this.motionZ *= 0.99;
-		
-		if(this.posY < 0)
-		{
-			setDead();
-		}
-		
-		int NUM = 5;
-		for(int i=0; i<NUM; i++)
-		{
-			double dx = (this.posX-this.prevPosX) / NUM;
-			double dy = (this.posY-this.prevPosY) / NUM;
-			double dz = (this.posZ-this.prevPosZ) / NUM;
-			FlansMod.proxy.spawnParticle("flansmod.fmflame",
-				this.prevPosX + dx*i,
-				this.prevPosY + dy*i,
-				this.prevPosZ + dz*i,
-				0,0,0);
-		}
-
 		if(this.onGround)
 		{
 			setDead();
