+++ src/main/java/com/flansmod/client/particle/EntitySmokeGrenade.java
@@ -1,4 +1,4 @@
-package com.flansmod.client;
+package com.flansmod.client.particle;
 
 import java.util.List;
 
@@ -22,16 +22,18 @@ import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.World;
 
-public class EntityFMMuzzleFlash extends EntityFX
+public class EntitySmokeGrenade extends EntityFX
 {
-	public EntityFMMuzzleFlash(World w, double px, double py, double pz, double mx, double my, double mz)
+	public int dischargeTime;
+	public EntitySmokeGrenade(World w, double px, double py, double pz, double mx, double my, double mz)
 	{
 		super(w, px, py, pz, mx, my, mz);
-		this.particleMaxAge *= 0.25;
-		this.particleGravity = 0;
+		this.particleMaxAge *= 20;
+		this.particleGravity = 1;
 		this.motionX = mx;
 		this.motionY = my;
 		this.motionZ = mz;
+		this.dischargeTime = 20;
 	}
 	
 	public int getFXLayer()
@@ -43,7 +45,7 @@ public class EntityFMMuzzleFlash extends EntityFX
 	{
 			return 1.0F;
 	}
-
+	
     public void renderParticle(Tessellator par1Tessellator, float par2, float par3, float par4, float par5, float par6, float par7)
     {
         //func_98187_b() = bindTexture();
@@ -51,21 +53,11 @@ public class EntityFMMuzzleFlash extends EntityFX
     	par1Tessellator.startDrawingQuads();
 		GL11.glAlphaFunc(GL11.GL_GREATER, 0.001F);
 		GL11.glEnable(GL11.GL_BLEND);
-
 		int srcBlend = GL11.glGetInteger(GL11.GL_BLEND_SRC);
 		int dstBlend = GL11.glGetInteger(GL11.GL_BLEND_DST);
-		GL11.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
-		GL11.glDepthMask(false);
-		double progress = (double)this.particleAge / (double)this.particleMaxAge;
-		if (progress < 0.3) {
-			FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "particle/Boom1.png"));
-		} else if (progress < 0.6) {
-			FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "particle/Boom2.png"));
-		} else if (progress < 0.8) {
-			FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "particle/Boom3alt.png"));
-		} else {
-			FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "particle/Boom4alt.png"));
-		}
+		GL11.glBlendFunc(1, GL11.GL_ONE_MINUS_SRC_ALPHA);
+		GL11.glDepthMask(false); 
+    	FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation("flansmod", "particle/case.png"));
 
         float scale = 0.1F * this.particleScale;
         float xPos = (float) (this.prevPosX + (this.posX - this.prevPosX) * (double) par2 - interpPosX);
@@ -73,15 +65,15 @@ public class EntityFMMuzzleFlash extends EntityFX
         float zPos = (float) (this.prevPosZ + (this.posZ - this.prevPosZ) * (double) par2 - interpPosZ);
         float colorIntensity = 1.0F;
         par1Tessellator.setColorOpaque_F(this.particleRed * colorIntensity, this.particleGreen * colorIntensity, this.particleBlue * colorIntensity);//, 1.0F);
-		par1Tessellator.setBrightness(15728880);
+
         par1Tessellator.addVertexWithUV((double) (xPos - par3 * scale - par6 * scale), (double) (yPos - par4 * scale), (double) (zPos - par5 * scale - par7 * scale), 0D, 1D);
         par1Tessellator.addVertexWithUV((double) (xPos - par3 * scale + par6 * scale), (double) (yPos + par4 * scale), (double) (zPos - par5 * scale + par7 * scale), 1D, 1D);
         par1Tessellator.addVertexWithUV((double) (xPos + par3 * scale + par6 * scale), (double) (yPos + par4 * scale), (double) (zPos + par5 * scale + par7 * scale), 1D, 0D);
         par1Tessellator.addVertexWithUV((double) (xPos + par3 * scale - par6 * scale), (double) (yPos - par4 * scale), (double) (zPos + par5 * scale - par7 * scale), 0D, 0D);
         par1Tessellator.draw();
-		GL11.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
+		GL11.glBlendFunc(srcBlend, dstBlend);
 		GL11.glDisable(GL11.GL_BLEND);
-		GL11.glDepthMask(true);
+		GL11.glDepthMask(true); 
 		GL11.glPopMatrix();
     }
 	
@@ -98,15 +90,60 @@ public class EntityFMMuzzleFlash extends EntityFX
 		this.prevPosY = this.posY;
 		this.prevPosZ = this.posZ;
 
-		this.motionX += rand.nextGaussian() * 0.005;
-		this.motionY += rand.nextGaussian() * 0.005;
-		this.motionZ += rand.nextGaussian() * 0.005;
-		this.moveEntity(this.motionX, this.motionY, this.motionZ);
-		this.motionX *= 0.5;
-		this.motionY *= 0.1;
-		this.motionZ *= 0.5;
+		if(this.particleAge++ >= this.particleMaxAge)
+		{
+			this.setDead();
+		}
+		
 
-		if(this.onGround || this.particleAge++ >= this.particleMaxAge)
+
+		this.motionY -= 0.04D * (double) this.particleGravity;
+		this.moveEntity(this.motionX, this.motionY, this.motionZ);
+		this.motionX *= 0.99;
+		this.motionY *= 0.99;
+		this.motionZ *= 0.99;
+		
+		if(worldObj.isAnyLiquid(this.boundingBox))
+		{
+			this.motionY = 1;
+			
+		}
+		
+		dischargeTime --;
+		
+		if(dischargeTime < 0)
+		{
+			double dx = (this.posX-this.prevPosX);
+			double dy = (this.posY-this.prevPosY);
+			double dz = (this.posZ-this.prevPosZ);
+			FlansMod.proxy.spawnParticle("flansmod.smokeburst",
+					this.posX,
+					this.posY,
+					this.posZ,
+					0,0,0);
+			
+			FlansMod.proxy.spawnParticle("flansmod.bigsmoke",
+					this.posX,
+					this.posY,
+					this.posZ,
+					0,0,0);
+			setDead();	
+		}
+		
+		int NUM = 5;
+		for(int i=0; i<NUM; i++)
+		{
+			double dx = (this.posX-this.prevPosX) / NUM;
+			double dy = (this.posY-this.prevPosY) / NUM;
+			double dz = (this.posZ-this.prevPosZ) / NUM;
+			FlansMod.proxy.spawnParticle("explode",
+				this.prevPosX + dx*i,
+				this.prevPosY + dy*i,
+				this.prevPosZ + dz*i,
+				0,0,0);
+		}
+
+		if(this.onGround)
 		{
 			setDead();
 		}
